Add ability to configure preferences - specifically ability to ignore last pdf position.
Background: the viewer remembers last opened page and zoom for each
opened pdf. When the same pdf is opened second timein the same browser,
it is opened at last read page. All this information is stored in local
storage.

A customer application needs to open pdf on the first page, regardless
of where it was opened last time.